refactor(telegram): use intent package types via re-exports (GH-810)
Update telegram adapter to use the shared intent package types:
- Replace local AnthropicClient with type alias to intent.AnthropicClient
- Replace local ConversationStore with type alias to intent.ConversationStore
- Replace local ConversationMessage with type alias to intent.ConversationMessage
- Move classifier and conversation tests to internal/intent package
- Remove duplicate Intent type and constants from classifier.go
- Add ConversationMessage and ClassifyResponse types to intent.go

This completes the Option A (type aliases) approach for minimal handler.go changes.
