+++ internal/intent/classifier_test.go
@@ -1,4 +1,4 @@
-package telegram
+package intent
 
 import (
 	"context"
